Move uptime cog to cogs folder
Pretty self-explanatory
